Inline chat widget tweaks (#206954)
* make status menu configurable

https://github.com/microsoft/vscode/issues/206940

* pull inline chat zone widget into its own file

* :lipstick:

* remove expand/crop feature

* make chat message contents scrollable
