+++ src/vs/workbench/contrib/inlineChat/browser/inlineChat.contribution.ts
@@ -51,8 +51,6 @@ registerAction2(InlineChatActions.FocusInlineChat);
 registerAction2(InlineChatActions.PreviousFromHistory);
 registerAction2(InlineChatActions.NextFromHistory);
 registerAction2(InlineChatActions.ViewInChatAction);
-registerAction2(InlineChatActions.ExpandMessageAction);
-registerAction2(InlineChatActions.ContractMessageAction);
 
 registerAction2(InlineChatActions.ToggleDiffForChange);
 registerAction2(InlineChatActions.FeebackHelpfulCommand);
